Fixing Trainer Delete

+++ GymManagment.BLL/Services/Classes/TrainerService.cs
@@ -110,6 +110,8 @@ public class TrainerService : ITrainerService
         var trainerRepo = _unitOfWork.GetRepository<Trainer>();
         var trainer = await trainerRepo.GetByIdAsync(id, ct);
         if (trainer == null) return ServiceResult.Failure("NotFound", "Trainer Not Found");
+        var hasCurrentOrFutureSessions = await _unitOfWork.GetRepository<Session>().ExistsAsync(s => s.TrainerId == trainer.Id && s.EndDate > DateTime.Now);
+        if (hasCurrentOrFutureSessions) return ServiceResult.Failure("TrainerId", "Cannot Delete Trainer With Ongoing Or Upcoming Sessions");
         
         trainerRepo.Delete(trainer);
         var result = await _unitOfWork.SaveChangesAsync(ct);
